requirements.txt
Flask
flask-cors
python-dotenv
SQLAlchemy
psycopg2-binary
gunicorn
